requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vapi-python==0.1.9
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pydantic==1.10.13
python-multipart==0.0.6
requests==2.31.0
websockets==12.0
streamlit==1.28.1
pandas==2.1.2
plotly==5.17.0